Replace 'element' XML element with name of metadata
Instead of using a generic 'element' XML element with a child
'name' XML element that contains the name of the metadata field,
we are going to create XML elements that use that name as the tag,
i.e <Title>, <FormGenre>, etc.
Note that the name of the field is Form/Genre, so we had to
remove the '/' before using the name as a tag.